convert callback-timing
